requirements.txt
amass==0.3.3
bitsandbytes==0.42.0
clip==1.0
flash_attn==2.5.7
h5py==3.11.0
joblib
jsonlines
loralib==0.1.2
matplotlib==3.9.3
nltk==3.8.1
numpy==1.23.0
omegaconf==2.3.0
openai
openrlhf==0.2.3
optimum==1.19.0
peft==0.10.0
Pillow==11.0.0
PyYAML
Requests
ring_flash_attn==0.1.1
rouge==1.0.1
scipy
smplx==0.1.28
spacy
tensorflow==2.18.0
torch==2.1.0+cu121
tqdm
transformers==4.37.2
trimesh==4.3.2
wandb==0.16.6
